==> scripts/16b_rangeShiftGeography.R <==
# Finding mean latitude of modeled ranges
# ~9 min per spp
startTime = Sys.time()

#### Import packages ----
library(raster)
library(tidyverse)
library(kableExtra)

#### Define parameters ----

# Species info
speciesList = c("Ovis canadensis",
                "Ovis dalli",
                "Oreamnos americanus",
                "Ovibos moschatus",
                "Bison bison")

#### Import data ----
sdmData = readRDS("models/sdm_10k/MaxEnt.RDS")

#### Define functions ----
extractLat = function(mySp) {
  message(mySp)
  message(Sys.time())
  maxEntSp = sdmData[[mySp]]
  ## Select the max sensitivity + specificity threshold
  maxEntThresh = maxEntSp$Thresholds[5,2]
  # Read prediction data
  distPreds_stack = raster(paste0("data/modeled_dist_10k/Predicted_Present_",
                                  mySp,
                                  ".tif"))
  distPreds_stack = projectRaster(distPreds_stack,
                                  crs = "+init=epsg:4326",
                                  method = "ngb")
  names(distPreds_stack) = "Predicted"
  # Identify scenarios
  mySSPs = c("45","85")
  # A function to compile all forecast distributions for a scenario
  makeSSPras = function(ssp){
    # Identify relevant files
    futureFiles = list.files(path = "data/modeled_dist_10k/",
                             pattern = "Future_",
                             full.names = TRUE)
    glmFutures = futureFiles[grep(pattern = mySp,
                                  x = futureFiles)]
    glmSSPfns = glmFutures[grep(pattern = ssp,
                                x = glmFutures)]
    ## Find the name of the cmip models
    cmipFileNames = basename(glmSSPfns) 
    cmipFileNames = gsub(pattern = "Predicted_Future_",
                         replacement = "",
                         x = cmipFileNames)
    cmipFileNames = gsub(pattern = paste0("_",mySp),
                         replacement = "",
                         x = cmipFileNames)
    cmipModelNames = gsub(pattern = ".tif",
                          replacement = "",
                          x = cmipFileNames)
    
    # Read prediction data
    distFuts_list = lapply(X = glmSSPfns,
                           FUN = raster)
    # Rename raster layers according to generation scheme in predictDist()
    distFuts_list <- lapply(distFuts_list,
                            FUN = function(x){
                              names(x) <- "Predicted"
                              return(x)})
    for(i in 1:length(distFuts_list)){
      names(distFuts_list)[[i]] <- paste0(names(distFuts_list[[i]]),
                                          "_",
                                          cmipModelNames[i])
    } 
    distFuts_stack = raster::stack(distFuts_list)
    distFuts_stack = projectRaster(distFuts_stack,
                                   crs = "+init=epsg:4326",
                                   method = "ngb")
  }
  
  ssp45Stack = makeSSPras(mySSPs[1])
  ssp85Stack = makeSSPras(mySSPs[2])
  
  # Reassign raster values based on Maxent threshold
  distPreds_stack[distPreds_stack < maxEntThresh] <- 0
  distPreds_stack[distPreds_stack >= maxEntThresh] <- 1
  ssp45Stack[ssp45Stack < maxEntThresh] <- 0
  ssp45Stack[ssp45Stack >= maxEntThresh] <- 1
  ssp85Stack[ssp85Stack < maxEntThresh] <- 0
  ssp85Stack[ssp85Stack >= maxEntThresh] <- 1
  
  ## Calculate range shift
  # https://stackoverflow.com/questions/52652444/calculating-centroid-of-raster
  currLatMean = colMeans(xyFromCell(distPreds_stack, 
                                    which(distPreds_stack[]==1)),
                         na.rm = T)[2]
  names(currLatMean) <- NULL
  
  ssp2LatMeans = list()
  for(i in 1:nlayers(ssp45Stack)){
    ssp2LatMeans[[i]] = colMeans(xyFromCell(ssp45Stack[[i]], 
                                       which(ssp45Stack[[i]][]==1)),
                                 na.rm = T)[2]
  }

  ssp5LatMeans = list()
  for(i in 1:nlayers(ssp85Stack)){
    ssp5LatMeans[[i]] = colMeans(xyFromCell(ssp85Stack[[i]], 
                                           which(ssp85Stack[[i]][]==1)),
                                 na.rm = T)[2]
  }
  
  ## Find mean, s.e. of futures
  ssp2mean = mean(unlist(ssp2LatMeans))
  ssp2se = plotrix::std.error(unlist(ssp2LatMeans))
  ssp5mean = mean(unlist(ssp5LatMeans))
  ssp5se = plotrix::std.error(unlist(ssp5LatMeans))

  return(c("Current" = currLatMean,
           "meanSSP2" = ssp2mean,
           "seSSP2" = ssp2se,
           "meanSSP5" = ssp5mean,
           "seSSP5" = ssp5se))
}


#### Implement functions ----
latList = lapply(X = speciesList,
                 FUN = extractLat)

#### Format results ----
## Convert to DF
latDF = do.call(rbind,
                latList) %>%
  as.data.frame()
## Add species column
latDF$Species = speciesList
## Create readable columns
latDF = latDF %>%
  mutate(Current = sprintf("%.2f",Current),
         SSP2_45 = paste0(sprintf("%.2f",meanSSP2),
                          " ± ",
                          sprintf("%.2f",seSSP2)),
         SSP5_85 = paste0(sprintf("%.2f",meanSSP5),
                          " ± ",
                          sprintf("%.2f",seSSP5))) %>%
  select(Species,
         Current,
         SSP2_45,
         SSP5_85)


## Convert to kable and export
latDF %>%
  rename("SSP2-4.5" = SSP2_45,
         "SSP5-8.5" = SSP5_85) %>%
  kable(table.attr = "style='width:40%;'") %>%
  kable_classic() %>%
  save_kable(file = "tables/predictedLatChange.png",
             zoom = 4)
